Move graph creators to their own package.
